luu thay doi

+++ BanHang001/demo/src/main/java/com/example/demo/Repository/KichCoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.Repository;
 import com.example.demo.Entity.KichCo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KichCoRepository extends JpaRepository<KichCo, Integer> {
+public interface KichCoRepository extends JpaRepository<KichCo, Long> {
 }
